refactor(datatable): consolidate ColumnsMenu into single Popover variant (LAM-1627 step 1)
Delete the DropdownMenu-based duplicate at infinite-datatable/ui/columns-menu and migrate every consumer to the Popover-based @/components/ui/columns-menu. Make panelConfig and columnActions optional so simple tables (datasets, queues, signal sub-tables, etc.) can render the menu without custom-column affordances.
