--- frontend/components/ui/infinite-datatable/ui/columns-menu.tsx
@@ -1,106 +1,0 @@
-import { closestCenter, DndContext, KeyboardSensor, PointerSensor, useSensor, useSensors } from "@dnd-kit/core";
-import { restrictToVerticalAxis } from "@dnd-kit/modifiers";
-import {
-  arrayMove,
-  SortableContext,
-  sortableKeyboardCoordinates,
-  verticalListSortingStrategy,
-} from "@dnd-kit/sortable";
-import { DropdownMenuSeparator } from "@radix-ui/react-dropdown-menu";
-import { ListRestart } from "lucide-react";
-import React from "react";
-import { useStore } from "zustand";
-
-import { Button } from "@/components/ui/button.tsx";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu.tsx";
-import { useDataTableStore } from "@/components/ui/infinite-datatable/model/datatable-store.tsx";
-import { ColumnsMenuItem } from "@/components/ui/infinite-datatable/ui/columns-menu-item.tsx";
-
-interface ColumnsMenuProps {
-  lockedColumns?: string[];
-  columnLabels?: { id: string; label: string }[];
-}
-
-export default function ColumnsMenu({ lockedColumns = [], columnLabels = [] }: ColumnsMenuProps) {
-  const store = useDataTableStore();
-  const { resetColumns, columnOrder, setColumnOrder, columnVisibility, setColumnVisibility } = useStore(
-    store,
-    (state) => ({
-      resetColumns: state.resetColumns,
-      columnOrder: state.columnOrder,
-      setColumnOrder: state.setColumnOrder,
-      columnVisibility: state.columnVisibility,
-      setColumnVisibility: state.setColumnVisibility,
-    })
-  );
-  const sensors = useSensors(
-    useSensor(PointerSensor),
-    useSensor(KeyboardSensor, {
-      coordinateGetter: sortableKeyboardCoordinates,
-    })
-  );
-  function handleDragEnd(event: any) {
-    const { active, over } = event;
-
-    if (active && over && active.id !== over.id) {
-      const oldIndex = columnOrder.indexOf(active.id);
-      const newIndex = columnOrder.indexOf(over.id);
-      if (oldIndex !== -1 && newIndex !== -1) {
-        const newColumnOrder = arrayMove(columnOrder, oldIndex, newIndex);
-        setColumnOrder(newColumnOrder);
-      }
-    }
-  }
-
-  function handleToggleVisibility(columnId: string) {
-    if (lockedColumns.includes(columnId)) return;
-
-    setColumnVisibility({
-      ...columnVisibility,
-      [columnId]: !(columnVisibility[columnId] !== false),
-    });
-  }
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button className="text-secondary-foreground focus:outline-0" icon="columns2" variant="outline">
-          Columns
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="start" className="relative min-w-32">
-        <DropdownMenuGroup>
-          <DndContext
-            sensors={sensors}
-            collisionDetection={closestCenter}
-            onDragEnd={handleDragEnd}
-            modifiers={[restrictToVerticalAxis]}
-          >
-            <SortableContext items={columnOrder} strategy={verticalListSortingStrategy}>
-              {columnOrder.map((columnId) => (
-                <ColumnsMenuItem
-                  key={columnId}
-                  id={columnId}
-                  label={columnLabels?.find((col) => col.id === columnId)?.label || columnId}
-                  isVisible={columnVisibility[columnId] !== false}
-                  isLocked={lockedColumns.includes(columnId)}
-                  onToggleVisibility={handleToggleVisibility}
-                />
-              ))}
-            </SortableContext>
-          </DndContext>
-        </DropdownMenuGroup>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem onClick={resetColumns}>
-          <ListRestart className="w-3.5 h-3.5 text-secondary-foreground" />
-          Reset columns
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
